Refactor: separate wallet struct from lib.rs
